+++ tox.ini
@@ -85,7 +85,7 @@ per-file-ignores =
 
 [pytest]
 junit_family = xunit2
-addopts = --doctest-modules -n auto -k test_regression
+addopts = --doctest-modules -n auto
 markers =
     slow: Tests that take a long time to run.
     wip: Tests that are work-in-progress.
